@@ -1,17 +1,18 @@
 name: dartleaf
 description: Dart API for leafletjs 1.6.0, using package:js. Mobile friendly interactive map for web applications, that can use various backends. 
-version: 0.1.4
+version: 1.1.4
 author: Plamen Todorov <androplam@gmail.com>
 repository: https://github.com/neterror/dart_leaflet.git
 homepage: https://github.com/neterror/dart_leaflet
 
 environment:
-  sdk: '>=2.4.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  js: ^0.6.1
+  js: ^0.6.3
 
 dev_dependencies:
-  build_runner: ^1.5.0
-  build_web_compilers: ^2.1.0
-  pedantic: ^1.7.0
+  build_runner: ^2.1.4
+  build_web_compilers: ^3.2.1
+  lints: ^1.0.1
+ 
