@@ -1,12 +1,13 @@
 import 'dart:convert';
 import 'dart:html';
 import 'package:dartleaf/dartleaf.dart';
+import 'package:js/js.dart';
 import 'draw.dart';
 
 /// Left click adds a point, right click removes point
 class DrawPolygons implements Draw {
   final LeafletMap _map;
-  CircleMarker _firstMark;
+  CircleMarker? _firstMark;
 
   final _featureCollection = <String, dynamic>{
     'type': 'FeatureCollection',
@@ -14,7 +15,7 @@ class DrawPolygons implements Draw {
   };
 
   Polyline _polyline = Polyline([]);
-  Polyline _dashLine = Polyline([]);
+  final Polyline _dashLine = Polyline([]);
   // when JS method returns an object, defined in the dart sources, its type is not the Dart type
   // but is NativeJavaScriptObject instead. For example  CircleMaker::getLatLng() returns LatLng, but
   // in our code it is NativeJavaScriptObject and if you pass the value to some other method that
@@ -46,11 +47,11 @@ class DrawPolygons implements Draw {
   String get geoJson => jsonEncode(_featureCollection);
 
   @override
-  set active(bool draw) {
-    if (draw) {
+  set active(bool? draw) {
+    if (draw!) {
       _map.dragging.disable();
-      _map.on(E.mousedown, _buttonDown);
-      _map.on(E.mousemove, _mouseMove);
+      _map.on(E.mousedown, allowInterop(_buttonDown));
+      _map.on(E.mousemove, allowInterop(_mouseMove));
 
       _polyline.addTo(_map);
       _dashLine.addTo(_map);
@@ -65,8 +66,8 @@ class DrawPolygons implements Draw {
     }
   }
 
-  _buttonDown(LeafletMouseEvent e) {
-    MouseEvent mouse = e.originalEvent;
+  void _buttonDown(LeafletMouseEvent e) {
+    MouseEvent mouse = e.originalEvent as MouseEvent;
     if (mouse.button == 0) {
       //left mouse button
       _addPoint(e);
@@ -75,19 +76,19 @@ class DrawPolygons implements Draw {
     }
   }
 
-  _addPoint(LeafletMouseEvent e) {
+  void _addPoint(LeafletMouseEvent e) {
     if (_ready) {
       _finish();
     } else {
       var node = CircleMarker(e.latlng, _normalNode);
-      if (_firstMark == null) _firstMark = node;
+      _firstMark ??= node;
       _points.add(node);
       _polyline.addLatLng(e.latlng);
       node.addTo(_map);
     }
   }
 
-  _delPoint(LeafletMouseEvent e) {
+  void _delPoint(LeafletMouseEvent e) {
     var last = _points.removeLast();
     last.remove(); //remove the marker from the layer
 
@@ -102,54 +103,39 @@ class DrawPolygons implements Draw {
     }
   }
 
-  _hideDashLine() {
+  void _hideDashLine() {
     _dashLine.setLatLngs([]);
     _dashLine.redraw();
   }
 
-  _drawDashLine(LeafletMouseEvent e) {
+  void _drawDashLine(LeafletMouseEvent e) {
     var last = _points.last.getLatLng();
     var line = [last, e.latlng];
     _dashLine.setLatLngs(line);
     _dashLine.redraw();
   }
 
-  _mouseMove(LeafletMouseEvent e) {
+  void _mouseMove(LeafletMouseEvent e) {
     if (_points.isEmpty) return;
     _drawDashLine(e);
-    double distance = _firstMark.getLatLng().distanceTo(e.latlng);
-    _ready = (distance < 20);
+    double distance = _firstMark!.getLatLng().distanceTo(e.latlng);
+    _ready = (distance < 40);
     if (_ready) {
-      _firstMark.setStyle(_accentNode);
+      _firstMark!.setStyle(_accentNode);
     } else {
-      _firstMark.setStyle(_normalNode);
+      _firstMark!.setStyle(_normalNode);
     }
   }
 
-  _finish() {
-    _polyline.addLatLng(_firstMark.getLatLng());
-    _firstMark.setStyle(_normalNode);
+  void _finish() {
+    _polyline.addLatLng(_firstMark!.getLatLng());
+    _firstMark!.setStyle(_normalNode);
     _hideDashLine();
-
-    _addFeature();
-
     _points.clear();
     _polyline = Polyline([]);
     _polyline.addTo(_map);
-
     _firstMark = null;
     _ready = false;
   }
 
-  void _addFeature() {
-    var geojson = _polyline.toGeoJSON();
-    var json = {
-      'type': 'feature',
-      'geometry': {
-        'type': geojson.geometry.type,
-        'coordinates': geojson.geometry.coordinates
-      }
-    };
-    _featureCollection['features'].add(json);
-  }
 }
