@@ -1,4 +1,4 @@
 abstract class Draw {
-  set active(bool draw);
+  set active(bool? draw);
   String get geoJson;
 }
