@@ -5,7 +5,7 @@ import 'package:js/js.dart';
 
 @JS()
 @anonymous
-mixin Evented {
+class Evented {
   /// Adds a listener function (fn) to a particular event type of the
   /// object. You can optionally specify the context of the listener
   /// (object the this keyword will point to). You can also pass several
@@ -16,13 +16,13 @@ mixin Evented {
   /// specified, it will remove all the listeners of that particular event
   /// from the object. Note that if you passed a custom context to on, you
   /// must pass the same context to off in order to remove the listener.
-  external Evented off(String type, [Function fn, dynamic context]);
+  external Evented off(String type, [Function? fn, dynamic context]);
 
   /// Fires an event of the specified type. You can optionally provide an
   /// data object — the first argument of the listener function will
   /// contain its properties. The event can optionally be propagated to
   /// event parents.
-  external Evented fire(String type, [dynamic data, bool propagate]);
+  external Evented fire(String type, [dynamic data, bool? propagate]);
 
   /// Returns true if a particular event type has any listeners attached to it.
   external bool listens(String type);
